feat: refactor service user profiles module + create missing mailchimp profile (#528)
