+++ src/event-logger/event-logger.module.ts
@@ -10,6 +10,7 @@ import { SubscriptionEntity } from 'src/entities/subscription.entity';
 import { TherapySessionEntity } from 'src/entities/therapy-session.entity';
 import { UserEntity } from 'src/entities/user.entity';
 import { PartnerAccessService } from 'src/partner-access/partner-access.service';
+import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
 import { SubscriptionUserService } from 'src/subscription-user/subscription-user.service';
 import { SubscriptionService } from 'src/subscription/subscription.service';
 import { TherapySessionService } from 'src/therapy-session/therapy-session.service';
@@ -33,6 +34,7 @@ import { EventLoggerService } from './event-logger.service';
   providers: [
     EventLoggerService,
     UserService,
+    ServiceUserProfilesService,
     SubscriptionUserService,
     TherapySessionService,
     PartnerAccessService,
