+++ src/service-user-profiles/service-user-profiles.service.spec.ts
@@ -1,3 +1,6 @@
+import { createMock } from '@golevelup/ts-jest';
+import { Test, TestingModule } from '@nestjs/testing';
+import { getRepositoryToken } from '@nestjs/typeorm';
 import {
   createCrispProfile,
   updateCrispProfile,
@@ -9,6 +12,7 @@ import {
   updateMailchimpProfile,
 } from 'src/api/mailchimp/mailchimp-api';
 import { UserEntity } from 'src/entities/user.entity';
+import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
 import {
   mockAltPartnerAccessEntity,
   mockCourseUserEntity,
@@ -16,34 +20,43 @@ import {
   mockPartnerEntity,
   mockUserEntity,
 } from 'test/utils/mockData';
+import { mockUserRepositoryMethods } from 'test/utils/mockedServices';
+import { Repository } from 'typeorm';
 import {
   EMAIL_REMINDERS_FREQUENCY,
   SIMPLYBOOK_ACTION_ENUM,
   mailchimpMarketingPermissionId,
-} from './constants';
-import {
-  createMailchimpCourseMergeField,
-  createServiceUserProfiles,
-  serializePartnersString,
-  serializeUserData,
-  updateServiceUserEmailAndProfiles,
-  updateServiceUserProfilesCourse,
-  updateServiceUserProfilesPartnerAccess,
-  updateServiceUserProfilesTherapy,
-  updateServiceUserProfilesUser,
-} from './serviceUserProfiles';
+} from '../utils/constants';
 
 jest.mock('src/api/crisp/crisp-api');
 jest.mock('src/api/mailchimp/mailchimp-api');
 
 describe('Service user profiles', () => {
+  let service: ServiceUserProfilesService;
+  const mockedUserRepository = createMock<Repository<UserEntity>>(mockUserRepositoryMethods);
+
   beforeEach(async () => {
     jest.clearAllMocks();
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [
+        ServiceUserProfilesService,
+        {
+          provide: getRepositoryToken(UserEntity),
+          useValue: mockedUserRepository,
+        },
+      ],
+    }).compile();
+
+    service = module.get<ServiceUserProfilesService>(ServiceUserProfilesService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
   });
 
   describe('createServiceUserProfiles', () => {
     it('should create crisp and mailchimp profiles for a public user', async () => {
-      await createServiceUserProfiles(mockUserEntity);
+      await service.createServiceUserProfiles(mockUserEntity);
 
       expect(createCrispProfile).toHaveBeenCalledWith({
         email: mockUserEntity.email,
@@ -96,7 +109,11 @@ describe('Service user profiles', () => {
     });
 
     it('should create crisp and mailchimp profiles for a partner user', async () => {
-      await createServiceUserProfiles(mockUserEntity, mockPartnerEntity, mockPartnerAccessEntity);
+      await service.createServiceUserProfiles(
+        mockUserEntity,
+        mockPartnerEntity,
+        mockPartnerAccessEntity,
+      );
 
       const partnerName = mockPartnerEntity.name.toLowerCase();
       const createdAt = mockUserEntity.createdAt.toISOString();
@@ -152,17 +169,23 @@ describe('Service user profiles', () => {
     it('should not propagate external api call errors', async () => {
       const mocked = jest.mocked(createCrispProfile);
       mocked.mockRejectedValue(new Error('Crisp API call failed'));
-      await expect(createServiceUserProfiles(mockUserEntity)).resolves.not.toThrow();
+      await expect(service.createServiceUserProfiles(mockUserEntity)).resolves.not.toThrow();
       mocked.mockReset();
     });
   });
 
   describe('updateServiceUserProfilesUser', () => {
     it('should update crisp and mailchimp profile user data', async () => {
-      await updateServiceUserProfilesUser(mockUserEntity, false, mockUserEntity.email);
+      await service.updateServiceUserProfilesUser(
+        mockUserEntity,
+        false,
+        false,
+        mockUserEntity.email,
+      );
 
       const lastActiveAt = mockUserEntity.lastActiveAt.toISOString();
 
+      expect(updateCrispProfile).toHaveBeenCalledTimes(1);
       expect(updateCrispProfile).toHaveBeenCalledWith(
         {
           marketing_permission: mockUserEntity.contactPermission,
@@ -173,6 +196,7 @@ describe('Service user profiles', () => {
         mockUserEntity.email,
       );
 
+      expect(updateMailchimpProfile).toHaveBeenCalledTimes(1);
       expect(updateMailchimpProfile).toHaveBeenCalledWith(
         {
           language: mockUserEntity.signUpLanguage,
@@ -202,7 +226,7 @@ describe('Service user profiles', () => {
       };
       const lastActiveAt = mockUserEntity.lastActiveAt.toISOString();
 
-      await updateServiceUserProfilesUser(mockUser, false, mockUser.email);
+      await service.updateServiceUserProfilesUser(mockUser, false, false, mockUser.email);
 
       expect(updateCrispProfile).toHaveBeenCalledWith(
         {
@@ -236,7 +260,12 @@ describe('Service user profiles', () => {
     });
 
     it('should additionally call crisp base profile update if required', async () => {
-      await updateServiceUserProfilesUser(mockUserEntity, true, mockUserEntity.email);
+      await service.updateServiceUserProfilesUser(
+        mockUserEntity,
+        true,
+        false,
+        mockUserEntity.email,
+      );
 
       expect(updateCrispProfile).toHaveBeenCalled();
       expect(updateMailchimpProfile).toHaveBeenCalled();
@@ -252,11 +281,41 @@ describe('Service user profiles', () => {
       );
     });
 
+    it("should update the user's email in crisp and mailchimp", async () => {
+      const oldEmail = mockUserEntity.email;
+      const newEmail = 'newemail@test.com';
+      await service.updateServiceUserProfilesUser(
+        { ...mockUserEntity, email: newEmail },
+        true,
+        true,
+        oldEmail,
+      );
+      const serialisedMockUserData = service.serializeUserData(mockUserEntity);
+      expect(updateCrispProfileBase).toHaveBeenCalledWith(
+        { email: newEmail, person: { locales: ['en'], nickname: 'name' } },
+        oldEmail,
+      );
+      expect(updateCrispProfile).toHaveBeenCalledTimes(1);
+      expect(updateCrispProfile).toHaveBeenCalledWith(
+        {
+          email_reminders_frequency: EMAIL_REMINDERS_FREQUENCY.TWO_MONTHS,
+          last_active_at: mockUserEntity.lastActiveAt.toISOString(),
+          marketing_permission: true,
+          service_emails_permission: true,
+        },
+        newEmail,
+      );
+      expect(updateMailchimpProfile).toHaveBeenCalledWith(
+        { ...serialisedMockUserData.mailchimpSchema, email_address: newEmail },
+        oldEmail,
+      );
+    });
+
     it('should not propagate external api call errors', async () => {
       const mocked = jest.mocked(updateMailchimpProfile);
       mocked.mockRejectedValue(new Error('Mailchimp API call failed'));
       await expect(
-        updateServiceUserProfilesUser(mockUserEntity, false, mockUserEntity.email),
+        service.updateServiceUserProfilesUser(mockUserEntity, false, false, mockUserEntity.email),
       ).resolves.not.toThrow();
       mocked.mockReset();
     });
@@ -264,7 +323,10 @@ describe('Service user profiles', () => {
 
   describe('updateServiceUserProfilesPartnerAccess', () => {
     it('should update crisp and mailchimp profile partner access data', async () => {
-      await updateServiceUserProfilesPartnerAccess([mockPartnerAccessEntity], mockUserEntity.email);
+      await service.updateServiceUserProfilesPartnerAccess(
+        [mockPartnerAccessEntity],
+        mockUserEntity.email,
+      );
 
       const partnerString = mockPartnerAccessEntity.partner.name.toLowerCase();
 
@@ -302,9 +364,9 @@ describe('Service user profiles', () => {
 
     it('should update crisp and mailchimp profile multiple partner accesses data', async () => {
       const partnerAccesses = [mockPartnerAccessEntity, mockAltPartnerAccessEntity];
-      await updateServiceUserProfilesPartnerAccess(partnerAccesses, mockUserEntity.email);
+      await service.updateServiceUserProfilesPartnerAccess(partnerAccesses, mockUserEntity.email);
 
-      const partnerString = serializePartnersString(partnerAccesses);
+      const partnerString = service.serializePartnersString(partnerAccesses);
 
       expect(updateCrispProfileBase).toHaveBeenCalledWith(
         {
@@ -342,7 +404,10 @@ describe('Service user profiles', () => {
       const mocked = jest.mocked(updateCrispProfile);
       mocked.mockRejectedValue(new Error('Crisp API call failed'));
       await expect(
-        updateServiceUserProfilesPartnerAccess([mockPartnerAccessEntity], mockUserEntity.email),
+        service.updateServiceUserProfilesPartnerAccess(
+          [mockPartnerAccessEntity],
+          mockUserEntity.email,
+        ),
       ).resolves.not.toThrow();
       mocked.mockReset();
     });
@@ -360,7 +425,7 @@ describe('Service user profiles', () => {
         },
       ];
 
-      await updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
+      await service.updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
 
       const firstTherapySessionAt = therapySession.startDateTime.toISOString();
       const nextTherapySessionAt = therapySession.startDateTime.toISOString();
@@ -394,7 +459,7 @@ describe('Service user profiles', () => {
     it('should update crisp and mailchimp profile combined therapy data for new booking', async () => {
       const partnerAccesses = [mockPartnerAccessEntity, mockAltPartnerAccessEntity];
 
-      await updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
+      await service.updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
 
       const firstTherapySessionAt =
         mockPartnerAccessEntity.therapySession[0].startDateTime.toISOString();
@@ -431,7 +496,7 @@ describe('Service user profiles', () => {
     it('should update crisp and mailchimp profile combined therapy data for updated booking', async () => {
       const partnerAccesses = [mockPartnerAccessEntity, mockAltPartnerAccessEntity];
 
-      await updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
+      await service.updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
 
       const firstTherapySessionAt =
         mockPartnerAccessEntity.therapySession[0].startDateTime.toISOString();
@@ -470,7 +535,7 @@ describe('Service user profiles', () => {
         SIMPLYBOOK_ACTION_ENUM.CANCELLED_BOOKING;
       const partnerAccesses = [mockPartnerAccessEntity, mockAltPartnerAccessEntity];
 
-      await updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
+      await service.updateServiceUserProfilesTherapy(partnerAccesses, mockUserEntity.email);
 
       const firstTherapySessionAt =
         mockPartnerAccessEntity.therapySession[0].startDateTime.toISOString();
@@ -506,7 +571,7 @@ describe('Service user profiles', () => {
       const mocked = jest.mocked(updateMailchimpProfile);
       mocked.mockRejectedValue(new Error('Mailchimp API call failed'));
       await expect(
-        updateServiceUserProfilesTherapy(
+        service.updateServiceUserProfilesTherapy(
           [mockPartnerAccessEntity, mockAltPartnerAccessEntity],
           mockUserEntity.email,
         ),
@@ -517,7 +582,7 @@ describe('Service user profiles', () => {
 
   describe('updateServiceUserProfilesCourse', () => {
     it('should update crisp and mailchimp profile course data', async () => {
-      await updateServiceUserProfilesCourse(mockCourseUserEntity, mockUserEntity.email);
+      await service.updateServiceUserProfilesCourse(mockCourseUserEntity, mockUserEntity.email);
 
       expect(updateCrispProfile).toHaveBeenCalledWith(
         {
@@ -542,7 +607,7 @@ describe('Service user profiles', () => {
       const mocked = jest.mocked(updateCrispProfile);
       mocked.mockRejectedValue(new Error('Crisp API call failed'));
       await expect(
-        updateServiceUserProfilesCourse(mockCourseUserEntity, mockUserEntity.email),
+        service.updateServiceUserProfilesCourse(mockCourseUserEntity, mockUserEntity.email),
       ).resolves.not.toThrow();
       mocked.mockReset();
     });
@@ -550,7 +615,7 @@ describe('Service user profiles', () => {
 
   describe('createMailchimpCourseMergeField', () => {
     it('should create mailchimp course merge field', async () => {
-      await createMailchimpCourseMergeField('Full course name');
+      await service.createMailchimpCourseMergeField('Full course name');
 
       expect(createMailchimpMergeField).toHaveBeenNthCalledWith(
         1,
@@ -564,47 +629,6 @@ describe('Service user profiles', () => {
         'C_FCN_S',
         'text',
       );
-    });
-  });
-  describe('updateServiceUserEmailAndProfiles', () => {
-    it("should update the user's email in crisp and mailchimp", async () => {
-      const oldEmail = mockUserEntity.email;
-      const newEmail = 'newemail@test.com';
-      await updateServiceUserEmailAndProfiles({ ...mockUserEntity, email: newEmail }, oldEmail);
-      const serialisedMockUserData = serializeUserData(mockUserEntity);
-      expect(updateCrispProfileBase).toHaveBeenCalledWith(
-        { email: newEmail, person: { locales: ['en'], nickname: 'name' } },
-        oldEmail,
-      );
-      expect(updateCrispProfile).toHaveBeenCalledWith(
-        { ...serialisedMockUserData.crispSchema },
-        newEmail,
-      );
-      expect(updateMailchimpProfile).toHaveBeenCalledWith(
-        { ...serialisedMockUserData.mailchimpSchema, email_address: newEmail },
-        oldEmail,
-      );
-    });
-    it('should not throw if request to Mailchimp API call fails', async () => {
-      const mocked = jest.mocked(updateMailchimpProfile);
-      mocked.mockRejectedValue(new Error('Mailchimp API call failed'));
-      const oldEmail = mockUserEntity.email;
-      const newEmail = 'newemail@test.com';
-
-      await expect(
-        updateServiceUserEmailAndProfiles({ ...mockUserEntity, email: newEmail }, oldEmail),
-      ).resolves.not.toThrow();
-      mocked.mockReset();
-    });
-    it('should not throw if request to Crisp API call  fails', async () => {
-      const mocked = jest.mocked(updateCrispProfileBase);
-      mocked.mockRejectedValue(new Error('Crisp API call failed'));
-      const oldEmail = mockUserEntity.email;
-      const newEmail = 'newemail@test.com';
-      await expect(
-        updateServiceUserEmailAndProfiles({ ...mockUserEntity, email: newEmail }, oldEmail),
-      ).resolves.not.toThrow();
-      mocked.mockReset();
     });
   });
 });
