+++ src/session-feedback/session-feedback.module.ts
@@ -11,6 +11,7 @@ import { SubscriptionEntity } from 'src/entities/subscription.entity';
 import { TherapySessionEntity } from 'src/entities/therapy-session.entity';
 import { UserEntity } from 'src/entities/user.entity';
 import { PartnerAccessService } from 'src/partner-access/partner-access.service';
+import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
 import { SessionService } from 'src/session/session.service';
 import { SubscriptionUserService } from 'src/subscription-user/subscription-user.service';
 import { SubscriptionService } from 'src/subscription/subscription.service';
@@ -37,6 +38,7 @@ import { SessionFeedbackService } from './session-feedback.service';
     SessionFeedbackService,
     SessionService,
     UserService,
+    ServiceUserProfilesService,
     SubscriptionUserService,
     SubscriptionService,
     PartnerAccessService,
