+++ src/session-user/session-user.module.ts
@@ -14,6 +14,7 @@ import { SubscriptionEntity } from 'src/entities/subscription.entity';
 import { TherapySessionEntity } from 'src/entities/therapy-session.entity';
 import { UserEntity } from 'src/entities/user.entity';
 import { PartnerService } from 'src/partner/partner.service';
+import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
 import { SubscriptionUserService } from 'src/subscription-user/subscription-user.service';
 import { SubscriptionService } from 'src/subscription/subscription.service';
 import { TherapySessionService } from 'src/therapy-session/therapy-session.service';
@@ -50,6 +51,7 @@ import { SessionUserService } from './session-user.service';
     PartnerAccessService,
     CourseService,
     PartnerService,
+    ServiceUserProfilesService,
     SubscriptionUserService,
     TherapySessionService,
     SubscriptionService,
