+++ src/user/user.module.ts
@@ -9,6 +9,7 @@ import { SubscriptionUserEntity } from 'src/entities/subscription-user.entity';
 import { SubscriptionEntity } from 'src/entities/subscription.entity';
 import { TherapySessionEntity } from 'src/entities/therapy-session.entity';
 import { UserEntity } from 'src/entities/user.entity';
+import { ServiceUserProfilesService } from 'src/service-user-profiles/service-user-profiles.service';
 import { SubscriptionUserService } from 'src/subscription-user/subscription-user.service';
 import { SubscriptionService } from 'src/subscription/subscription.service';
 import { TherapySessionService } from 'src/therapy-session/therapy-session.service';
@@ -36,6 +37,7 @@ import { UserService } from './user.service';
     UserService,
     AuthService,
     PartnerAccessService,
+    ServiceUserProfilesService,
     SubscriptionService,
     SubscriptionUserService,
     TherapySessionService,
